Tests/LibWeb/Text/expected/image-decode-resolves-after-state-is-completely-available.txt

width=120 height=120 naturalWidth=120 naturalHeight=120
